pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "gtkwedit"
version = "0.0.1"
dependencies = [
  "typing-extensions",
  "pyvcd"
]

[tool.isort]
profile = "black"
